\name{Crambled-package}
\alias{Crambled-package}
\alias{Crambled}
\docType{package}
\title{
Enabling intuitive resolution of conflicting cellularity estimates.
}
\description{
This package combines the scripts and shiny App associated with the paper "Crambled: A Shiny application to enable intuitive resolution of conflicting cellularity estimates" (see references).
}
\details{
\tabular{ll}{
Package: \tab Crambled\cr
Type: \tab Package\cr
Version: \tab 2.0\cr
Date: \tab 2016-01-04\cr
License: \tab GPL-2\cr
}
See the referenced material for a guide on using the package
}
\author{
Andy Lynch
}
\references{
Lynch A. Crambled: A Shiny application to enable intuitive resolution of conflicting cellularity estimates F1000Research 2015, 4:1407 (doi: 10.12688/f1000research.7453.1)

https://github.com/dralynch/crambled}
\keyword{ package }