\name{RunCrambledShinyApp}
\alias{RunCrambledShinyApp}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Function to start the Crambled Shiny App}
\description{
Initiates the Crambled Shiny Application
}
\usage{
RunCrambledShinyApp()
}
\references{
Lynch A. Crambled: A Shiny application to enable intuitive resolution of conflicting cellularity estimates F1000Research 2015, 4:1407 (doi: 10.12688/f1000research.7453.1)

https://github.com/dralynch/crambled
}
\author{
Andy Lynch
}
\seealso{
\code{\link{CrambledScan}}
}
\examples{

## RunCrambledShinyApp()

}